v2.0.10,尝试更新生成mesh_evaluator的生成临时文件目录

+++ src/config/config.py
@@ -252,8 +252,8 @@ class UnifiedParameterSpace:
             # 网格质量参数
             'distortion_distance': ParameterDefinition(
                 name='distortion_distance',
-                param_type=ParameterType.INTEGER,
-                bounds=(10, 30),
+                param_type=ParameterType.FLOAT,
+                bounds=(10.0, 30.0),
                 description='扭曲距离',
                 unit='%',
                 ansa_mapping='distortion-distance',
